ager-productos-app/requirements.txt
django==4.0
djangorestframework==3.12.4
psycopg2==2.9.2
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers==3.10.1
pytz==2021.3